Crate aggregator_common

Structs

ThresholdAggregator
Generic data structure that combines data values and counts the number of provided data samples. It can also reveal an aggregated value only when there are enough data samples (more or equal to sample_threshold).

Traits

Monoid
Represents a data structure with a single associative binary operation (combine) and an identity element. https://en.wikipedia.org/wiki/Monoid